rename: oauth 도메인 테스트 디렉토리 구조 수정 #33
